requirements.txt
# Core dependencies
yfinance>=0.2.31
pandas>=2.1.0
numpy>=1.25.0

# Export / reporting
openpyxl>=3.1.2        # Excel export
tabulate>=0.9.0        # Pretty-print tables in terminal/reports

# Configuration
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0